intermediary/requirement.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
python-dotenv==0.13.0
requests==2.24.0
